refactor: 랜더 핸들러 분리
