+++ src/components/restaurant/Restaurant.ts
@@ -1,22 +1,7 @@
 import { Irestaurant } from "../../types";
 
-import categoryMatchedImageData from "./categoryMatchedImageData";
-import { baseTemplate, categoryTemplate, template } from "./template";
-
-const findCategory = (restaurant: Irestaurant) =>
-  categoryMatchedImageData.find(
-    (item) => item.category === restaurant.category,
-  );
-
-const getMatchedCategoryInfo = (restaurant: Irestaurant) => {
-  const categoryInfo = findCategory(restaurant);
-  if (categoryInfo) return categoryInfo;
-
-  return {
-    categoryImg: "",
-    category: "",
-  };
-};
+import { getMatchedCategoryInfo } from "./renderHandlers";
+import { baseTemplate, categoryTemplate, template } from "./restaurantTemplate";
 
 function Restaurant(restaurant: Irestaurant) {
   const categoryInfo = getMatchedCategoryInfo(restaurant);
